+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## **[1.4.0]** 19/02/2021
 ### **Added**
 - Add simplest ray tracer shader
+- Shiny cubes in mod space shader
 
 ## **[1.4.0]** 18/02/2021
 ### **Added**